+++ src/feature_info.rs
@@ -57,10 +57,8 @@ where
 fn get_shared_library_paths() -> Vec<Vec<wchar_t>> {
     let mut shared_library_paths = vec![];
 
-    #[cfg(not(target_os = "windows"))]
-    let platform = "Linux_x86_64";
-    #[cfg(target_os = "windows")]
-    let platform = "Windows_x86_64";
+    // Set by build.rs
+    let platform = env!("DLSS_SDK_PLATFORM");
 
     #[cfg(feature = "debug_overlay")]
     let profile = "dev";