+++ src/wrapper.h
@@ -1,5 +1,5 @@
 #include <vulkan/vulkan.h>
 #include <nvsdk_ngx_helpers.h>
-#include <nvsdk_ngx_helpers_dlssd.h>
+#include <nvsdk_ngx_helpers_dlssd_d3d.h>
 #include <nvsdk_ngx_helpers_vk.h>
 #include <nvsdk_ngx_helpers_dlssd_vk.h>